+++ packages/keystatic/DocumentEditor/code-block/code-block-ui.tsx
@@ -14,7 +14,6 @@ import { css, tokenSchema } from '@voussoir/style';
 import { Tooltip, TooltipTrigger } from '@voussoir/tooltip';
 import { Kbd, Text } from '@voussoir/typography';
 
-import Prism from '../prism';
 import { useToolbarState } from '../toolbar-state';
 import {
   BlockPopover,
@@ -22,6 +21,13 @@ import {
   BlockWrapper,
   ToolbarSeparator,
 } from '../primitives';
+import {
+  aliasesToCanonicalName,
+  aliasesToLabel,
+  canonicalNameToLabel,
+  labelToCanonicalName,
+  languagesWithAliases,
+} from './languages';
 
 function CodeButton() {
   const {
@@ -167,7 +173,7 @@ export function CodeElement({
                       { at: path }
                     );
                   }
-                } else {
+                } else if (typeof selection === 'string') {
                   if (selection === 'plain') {
                     Transforms.unsetNodes(editor, 'language', { at: path });
                     setInputValue('Plain text');
@@ -178,9 +184,7 @@ export function CodeElement({
                     { language: selection as string },
                     { at: path }
                   );
-                  const label = languages.find(
-                    lang => lang.value === selection
-                  )?.label;
+                  const label = canonicalNameToLabel.get(selection);
                   if (label) {
                     setInputValue(label);
                   }
@@ -224,78 +228,3 @@ export function CodeElement({
     </BlockWrapper>
   );
 }
-
-const languages = [
-  { label: 'C', value: 'c' },
-  { label: 'C++', value: 'cpp' },
-  { label: 'Arduino', value: 'arduino' },
-  { label: 'Bash', value: 'bash' },
-  { label: 'C#', value: 'csharp' },
-  { label: 'CSS', value: 'css' },
-  { label: 'Diff', value: 'diff' },
-  { label: 'Go', value: 'go' },
-  { label: 'INI', value: 'ini' },
-  { label: 'Java', value: 'java' },
-  { label: 'JavaScript', value: 'javascript' },
-  { label: 'JSX', value: 'jsx' },
-  { label: 'JSON', value: 'json' },
-  { label: 'Kotlin', value: 'kotlin' },
-  { label: 'Less', value: 'less' },
-  { label: 'Lua', value: 'lua' },
-  { label: 'Makefile', value: 'makefile' },
-  { label: 'Markdown', value: 'markdown' },
-  { label: 'Objective-C', value: 'objectivec' },
-  { label: 'Perl', value: 'perl' },
-  { label: 'PHP', value: 'php' },
-  { label: 'Python', value: 'python' },
-  { label: 'R', value: 'r' },
-  { label: 'Ruby', value: 'ruby' },
-  { label: 'Rust', value: 'rust' },
-  { label: 'Sass', value: 'sass' },
-  { label: 'SCSS', value: 'scss' },
-  { label: 'SQL', value: 'sql' },
-  { label: 'Swift', value: 'swift' },
-  { label: 'TypeScript', value: 'typescript' },
-  { label: 'TSX', value: 'tsx' },
-  { label: 'VB.NET', value: 'vbnet' },
-  { label: 'YAML', value: 'yaml' },
-];
-
-const canonicalNameToLabel = new Map(languages.map(x => [x.value, x.label]));
-const labelToCanonicalName = new Map(languages.map(x => [x.label, x.value]));
-
-const languageToCanonicalName = new Map(
-  languages.map(lang => [Prism.languages[lang.value], lang.value])
-);
-
-const aliasesToCanonicalName = new Map(
-  Object.keys(Prism.languages).flatMap(lang => {
-    const canonicalName = languageToCanonicalName.get(Prism.languages[lang]);
-    if (canonicalName === undefined) {
-      return [];
-    }
-    return [[lang, canonicalName]];
-  })
-);
-
-const languagesToAliases = new Map(
-  languages.map(lang => [lang.value, [] as string[]])
-);
-
-for (const [alias, canonicalName] of aliasesToCanonicalName) {
-  languagesToAliases.get(canonicalName)!.push(alias);
-}
-const languagesWithAliases = [
-  { label: 'Plain text', value: 'plain', aliases: [] },
-  ...[...languagesToAliases].map(([canonicalName, aliases]) => ({
-    label: canonicalNameToLabel.get(canonicalName)!,
-    value: canonicalName,
-    aliases,
-  })),
-];
-const aliasesToLabel = new Map(
-  [...aliasesToCanonicalName].map(([alias, canonicalName]) => [
-    alias,
-    canonicalNameToLabel.get(canonicalName)!,
-  ])
-);