+++ apps/remix/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @keystatic/remix-test-app
 
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies [d978f9b]
+  - @keystatic/core@0.0.85
+
+## 0.0.47
+
+### Patch Changes
+
+- Updated dependencies [a45bad6]
+  - @keystatic/core@0.0.84
+
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies [c7da5cd]
+- Updated dependencies [323d8cb]
+- Updated dependencies [6b9c653]
+- Updated dependencies [70116b6]
+  - @keystatic/core@0.0.83
+
 ## 0.0.45
 
 ### Patch Changes