+++ keystatic/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @keystatic/test-next-app
 
+## 0.0.55
+
+### Patch Changes
+
+- Updated dependencies [d978f9b]
+  - @keystatic/core@0.0.85
+
+## 0.0.54
+
+### Patch Changes
+
+- Updated dependencies [a45bad6]
+  - @keystatic/core@0.0.84
+
+## 0.0.53
+
+### Patch Changes
+
+- Updated dependencies [c7da5cd]
+- Updated dependencies [323d8cb]
+- Updated dependencies [6b9c653]
+- Updated dependencies [c0a4d82]
+- Updated dependencies [70116b6]
+  - @keystatic/core@0.0.83
+  - @voussoir/typography@0.1.3
+
 ## 0.0.52
 
 ### Patch Changes